refactor: add package for game entities
